+++ src/elements/conditions/OwnerSectionConditionRule.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace benf\neo\elements\conditions;
+
+use Craft;
+use craft\base\ElementInterface;
+use craft\elements\conditions\entries\SectionConditionRule;
+use craft\elements\Entry;
+
+/**
+ * Class OwnerSectionConditionRule
+ *
+ * @package benf\neo\elements\conditions
+ * @author Spicy Web <plugins@spicyweb.com.au>
+ * @since 3.4.0
+ */
+class OwnerSectionConditionRule extends SectionConditionRule
+{
+    use OwnerConditionRuleTrait;
+
+    /**
+     * @inheritdoc
+     */
+    public function getLabel(): string
+    {
+        return Craft::t('neo', 'Owner Section');
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function matchElement(ElementInterface $element): bool
+    {
+        $owner = $element->getOwner();
+        return !($owner instanceof Entry) || parent::matchElement($owner);
+    }
+}